Splitting the server code
