+++ deploy/docker/cega/mq/rabbitmq.conf
@@ -1,0 +1,3 @@
+disk_free_limit.absolute = 1GB
+management.listener.port = 15672
+management.load_definitions = /etc/rabbitmq/definitions.json
