+++ deploy/docker/cega/server/users.py
@@ -1,0 +1,42 @@
+import logging
+
+from base64 import b64decode
+import json
+
+from aiohttp import web
+
+LOG = logging.getLogger(__name__)
+
+USERS = {}
+HTTP_AUTH_USERNAME = 'fega'
+HTTP_AUTH_PASSWORD = 'testing' # yup, we don't care, it's just for testing
+
+async def get_user(request):
+
+    # Authenticate
+    auth_header = request.headers.get('AUTHORIZATION')
+    if not auth_header:
+        raise web.HTTPUnauthorized(reason='Protected access')
+    _, token = auth_header.split(None, 1)  # Skipping the Basic keyword
+    auth_user, auth_password = b64decode(token).decode().split(':', 1)
+    if HTTP_AUTH_USERNAME != auth_user or HTTP_AUTH_PASSWORD != auth_password:
+        raise web.HTTPUnauthorized(reason='Protected access')
+
+    # Search
+    term = request.match_info.get('term')
+    record = USERS.get(term)
+
+    if not record:
+        raise web.HTTPNotFound(reason='User not found')
+
+    return web.json_response(record,
+                             headers = { "Server": "Central EGA (test) Server",
+                                         "X-EGA-apiVersion" : "v2",
+                                         "X-EGA-docLink" : "https://ega-archive.org",
+                                        })
+
+def load_users(filepath):
+    global USERS
+    with open(filepath, 'r') as stream:
+        USERS = json.load(stream)
+    LOG.debug('Loaded %d users: %s', len(USERS) / 2, list(USERS.keys()))
